# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Tiny TODO API
  description: Very small API for managing todos — suitable for demo and docs practice.
  version: "0.1.0"
servers:
  - url: http://localhost:3000
paths:
  /todos:
    get:
      summary: List all todos
      responses:
        "200":
          description: Array of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
    post:
      summary: Create a todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTodo"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
  /todos/{id}:
    parameters:
      - $ref: "#/components/parameters/todoId"
    get:
      summary: Get a todo by id
      responses:
        "200":
          description: Todo found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "404":
          description: Not found
    put:
      summary: Update a todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTodo"
      responses:
        "200":
          description: Updated todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "404":
          description: Not found
    delete:
      summary: Delete a todo
      responses:
        "204":
          description: Deleted
components:
  parameters:
    todoId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: ID of the todo
  schemas:
    Todo:
      type: object
      properties:
        id:
          type: string
          example: "1"
        title:
          type: string
          example: "Buy milk"
        completed:
          type: boolean
          example: false
    NewTodo:
      type: object
      required: [title]
      properties:
        title:
          type: string
          example: "Buy milk"
    UpdateTodo:
      type: object
      properties:
        title:
          type: string
        completed:
          type: boolean